+++ Libraries/Lists/FlatList.js.flow
@@ -15,7 +15,15 @@ import typeof ScrollViewNativeComponent from '../Components/ScrollView/ScrollVie
 import {type ScrollResponderType} from '../Components/ScrollView/ScrollView';
 import type {ViewStyleProp} from '../StyleSheet/StyleSheet';
 import type {RenderItemType} from './VirtualizedList';
-import typeof VirtualizedList from './VirtualizedList';
+import type {
+  ViewToken,
+  ViewabilityConfigCallbackPair,
+} from '@react-native/virtualized-lists/Lists/ViewabilityHelper';
+import type {
+  RenderItemType,
+  RenderItemProps,
+} from '@react-native/virtualized-lists/Lists/VirtualizedList';
+import typeof VirtualizedList from '@react-native/virtualized-lists/Lists/VirtualizedList';
 
 type RequiredProps<ItemT> = {|
   /**
