+++ Libraries/Utilities/ReactNativeTestTools.js
@@ -15,8 +15,8 @@ import type {ReactTestRenderer as ReactTestRendererType} from 'react-test-render
 const Switch = require('../Components/Switch/Switch').default;
 const TextInput = require('../Components/TextInput/TextInput');
 const View = require('../Components/View/View');
-const VirtualizedList = require('../Lists/VirtualizedList').default;
 const Text = require('../Text/Text');
+const {VirtualizedList} = require('@react-native/virtualized-lists');
 const React = require('react');
 const ShallowRenderer = require('react-shallow-renderer');
 const ReactTestRenderer = require('react-test-renderer');
