+++ package.json
@@ -118,6 +118,7 @@
     "@react-native/gradle-plugin": "^0.72.2",
     "@react-native/js-polyfills": "^0.72.0",
     "@react-native/normalize-colors": "^0.72.0",
+    "@react-native/virtualized-lists": "1.0.0",
     "abort-controller": "^3.0.0",
     "anser": "^1.4.9",
     "base64-js": "^1.1.2",
