+++ types/__typetests__/index.tsx
@@ -57,7 +57,6 @@ import {
   KeyboardAvoidingView,
   LayoutChangeEvent,
   Linking,
-  ListRenderItemInfo,
   LogBox,
   Modal,
   MouseEvent,
@@ -108,7 +107,6 @@ import {
   UIManager,
   View,
   ViewStyle,
-  VirtualizedList,
   YellowBox,
   findNodeHandle,
   requireNativeComponent,
@@ -1847,32 +1845,6 @@ const DarkMode = () => {
   }, [color, subscription]);
 
   return <Text>Is dark mode enabled? {isDarkMode}</Text>;
-};
-
-// VirtualizedList
-// Test inspired by: https://reactnative.dev/docs/virtualizedlist
-const VirtualizedListTest = () => {
-  const DATA = [1, 2, 3];
-
-  const getItem = (data: number[], index: number) => {
-    return {
-      title: `Item ${data[index]}`,
-    };
-  };
-
-  const getItemCount = (data: number[]) => data.length;
-
-  return (
-    <VirtualizedList
-      data={DATA}
-      initialNumToRender={4}
-      renderItem={({item}: ListRenderItemInfo<ReturnType<typeof getItem>>) => (
-        <Text>{item.title}</Text>
-      )}
-      getItemCount={getItemCount}
-      getItem={getItem}
-    />
-  );
 };
 
 // DevSettings
