+++ types/public/DeprecatedPropertiesAlias.d.ts
@@ -37,7 +37,6 @@ import {
   ImagePropsAndroid,
   ImageBackgroundProps,
   FlatListProps,
-  VirtualizedListProps,
   SectionListProps,
   ModalProps,
   TouchableWithoutFeedbackProps,
@@ -144,9 +143,6 @@ export type ImageBackgroundProperties = ImageBackgroundProps;
 
 /** @deprecated Use FlatListProps */
 export type FlatListProperties<ItemT> = FlatListProps<ItemT>;
-
-/** @deprecated Use VirtualizedListProps */
-export type VirtualizedListProperties<ItemT> = VirtualizedListProps<ItemT>;
 
 /** @deprecated Use SectionListProps */
 export type SectionListProperties<ItemT> = SectionListProps<ItemT>;
